My last attempt at making SwiftUIFlux work for this
Something is really slow here, and I can't find what. The `Toggle` takes
like a second to actually flip. I think I'm doing the actions and updates
idiomatically, and this diff takes the DB access off the critical path
to update the state, but it's still slow.

Now I'm leaning towards just blaming the framework. I might rip it out
because SwiftUI already has most of the primitives we need, and the Flux
architecture is just adding overhead.